+++ app/archiver/mholt/mholt.go
@@ -1,0 +1,16 @@
+package mholt
+
+import (
+	"khranity/app/log"
+)
+
+type Provider struct {
+	logger 	*log.Logger
+}
+
+// NewProvider ...
+func NewProvider(logger *log.Logger) *Provider {
+	return &Provider{
+		logger: logger,
+	}
+}