+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2024-03-18
+* new: add ignore params to lore.yml
+* new: add various archivers (native/targz)
+* fix: error "archive/tar: write too long"
+* fix: some epic and awesome bugs
+
 ## 2024-03-15
 * upd: get ready for go version 1.22.1
 * new: add support BIG data